requirements.txt
torch>=1.9.0
transformers>=4.20.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0
joblib>=1.0.0
